+++ gamemodes/terrortown/gamemode/shared/sh_shopeditor.lua
@@ -2,43 +2,133 @@
 -- @author Alf21
 -- @module ShopEditor
 
+
 ShopEditor = ShopEditor or {}
+
+ShopEditor.MODE_DEFAULT = 1
+ShopEditor.MODE_ADDED = 2
+ShopEditor.MODE_INHERIT_ADDED = 3
+ShopEditor.MODE_INHERIT_REMOVED = 4
+
 ShopEditor.savingKeys = {
-	credits = {typ = "number", bits = 8, default = 1}, -- from 0 to 255 (2^8 - 1)
-	minPlayers = {typ = "number", bits = 6}, -- from 0 to 63 (2^6 - 1)
-	globalLimited = {typ = "bool"}, -- 0 and 1
-	limited = {typ = "bool"}, -- 0 and 1
-	NoRandom = {typ = "bool"}, -- 0 and 1
-	notBuyable = {typ = "bool"}, -- 0 and 1
-	teamLimited = {typ = "bool"} -- 0 and 1
+	notBuyable = {
+		order = 1,
+		typ = "bool",
+		default = false,
+		name = "not_buyable",
+		inverted = true,
+		b_desc = true
+	},
+	NoRandom = {
+		order = 2,
+		typ = "bool",
+		default = false,
+		name = "not_random",
+		inverted = false,
+		b_desc = true
+	},
+	globalLimited = {
+		order = 3,
+		typ = "bool",
+		default = false,
+		name = "global_limited",
+		inverted = false,
+		b_desc = true
+	},
+	teamLimited = {
+		order = 4,
+		typ = "bool",
+		default = false,
+		name = "team_limited",
+		inverted = false,
+		b_desc = true
+	},
+	limited = {
+		order = 5,
+		typ = "bool",
+		default = false,
+		name = "player_limited",
+		inverted = false,
+		b_desc = true
+	},
+	minPlayers = {
+		order = 6,
+		typ = "number",
+		bits = 6,
+		default = 0,
+		min = 0,
+		max = 63,
+		name = "min_players",
+		b_desc = false
+	},
+	credits = {
+		order = 7,
+		typ = "number",
+		bits = 5,
+		default = 1,
+		min = 1,
+		max = 20,
+		name = "credits",
+		b_desc = false
+	}
 }
+
+ShopEditor.savingKeysCount = table.Count(ShopEditor.savingKeys)
+ShopEditor.savingKeysBitCount = math.ceil(math.log(ShopEditor.savingKeysCount, 2))
 
 ShopEditor.cvars = {
 	ttt2_random_shops = {
-		order = 1, typ = "bool", default = 0,
-		name = "random_shops", b_desc = true
+		order = 1,
+		typ = "bool",
+		default = 0,
+		name = "random_shops",
+		b_desc = true
 	},
 	ttt2_random_shop_items = {
-		order = 2, typ = "number", bits = 6, default = 10, min = 1, max = 60,
-		name = "random_shop_items", b_desc = true
+		order = 2,
+		typ = "number",
+		bits = 6,
+		default = 10,
+		min = 1,
+		max = 60,
+		name = "random_shop_items",
+		b_desc = true
 	},
 	ttt2_random_team_shops = {
-		order = 3, typ = "bool", default = 1,
-		name = "random_team_shops", b_desc = false
+		order = 3,
+		typ = "bool",
+		default = 1,
+		name = "random_team_shops",
+		b_desc = false
 	},
 	ttt2_random_shop_reroll = {
-		order = 4, typ = "bool", default = 1,
-		name = "random_shop_reroll", b_desc = false
+		order = 4,
+		typ = "bool",
+		default = 1,
+		name = "random_shop_reroll",
+		b_desc = false
 	},
 	ttt2_random_shop_reroll_cost = {
-		order = 5, typ = "number", bits = 4, default = 1, min = 0, max = 10,
-		name = "random_shop_reroll_cost", b_desc = false
+		order = 5,
+		typ = "number",
+		bits = 4,
+		default = 1,
+		min = 0,
+		max = 10,
+		name = "random_shop_reroll_cost",
+		b_desc = false
 	},
 	ttt2_random_shop_reroll_per_buy = {
-		order = 6, typ = "bool", default = 0,
-		name = "random_shop_reroll_per_buy", b_desc = false
+		order = 6,
+		typ = "bool",
+		default = 0,
+		name = "random_shop_reroll_per_buy",
+		b_desc = false
 	}
 }
+
+-- Table which contains all equipment and is sorted by their translated equipment names
+ShopEditor.sortedEquipmentList = {}
 
 local net = net
 local pairs = pairs
@@ -50,6 +140,8 @@ local pairs = pairs
 function ShopEditor.InitDefaultData(item)
 	if not item then return end
 
+	item.defaultValues = item.defaultValues or {}
+
 	for key, data in pairs(ShopEditor.savingKeys) do
 		if item[key] == nil then
 			if data.typ == "number" then
@@ -60,6 +152,8 @@ function ShopEditor.InitDefaultData(item)
 				item[key] = data.default or ""
 			end
 		end
+
+		item.defaultValues[key] = item[key]
 	end
 end
 
@@ -77,8 +171,11 @@ function ShopEditor.WriteItemData(messageName, name, item, plys)
 
 	net.Start(messageName)
 	net.WriteString(name)
+	net.WriteUInt(ShopEditor.savingKeysCount, ShopEditor.savingKeysBitCount or 16)
 
 	for key, data in pairs(ShopEditor.savingKeys) do
+		net.WriteString(key)
+
 		if data.typ == "number" then
 			net.WriteUInt(item[key], data.bits or 16)
 		elseif data.typ == "bool" then
@@ -123,7 +220,12 @@ function ShopEditor.ReadItemData()
 		return name
 	end
 
-	for key, data in pairs(ShopEditor.savingKeys) do
+	local keyCount = net.ReadUInt(ShopEditor.savingKeysBitCount or 16)
+
+	for i = 1, keyCount do
+		local key = net.ReadString()
+		local data = ShopEditor.savingKeys[key]
+
 		if data.typ == "number" then
 			equip[key] = net.ReadUInt(data.bits or 16)
 		elseif data.typ == "bool" then