+++ lua/terrortown/menus/gamemode/administration.lua
@@ -1,0 +1,12 @@
+--- @ignore
+
+CLGAMEMODEMENU.base = "base_gamemodemenu"
+
+CLGAMEMODEMENU.icon = Material("vgui/ttt/vskin/helpscreen/administration")
+CLGAMEMODEMENU.title = "menu_administration_title"
+CLGAMEMODEMENU.description = "menu_administration_description"
+CLGAMEMODEMENU.priority = 50
+
+function CLGAMEMODEMENU:IsAdminMenu()
+	return true
+end